Fix flaky tests
