+++ src/test/java/land/oras/RegistryWireMockTest.java
@@ -25,16 +25,32 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
+import land.oras.auth.FileStoreAuthenticationProvider;
+import land.oras.auth.UsernamePasswordProvider;
+import land.oras.credentials.FileStore;
 import land.oras.utils.JsonUtils;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @WireMockTest
+@Execution(ExecutionMode.SAME_THREAD)
 public class RegistryWireMockTest {
 
     private static final Logger LOG = LoggerFactory.getLogger(RegistryWireMockTest.class);
+
+    private final UsernamePasswordProvider authProvider = new UsernamePasswordProvider("myuser", "mypass");
+
+    @TempDir
+    private Path configDir;
 
     @Test
     void shouldListTags(WireMockRuntimeInfo wmRuntimeInfo) {
@@ -45,10 +61,57 @@ public class RegistryWireMockTest {
                 .willReturn(WireMock.okJson(JsonUtils.toJson(new Tags("artifact-text", List.of("latest", "0.1.1"))))));
 
         // Insecure registry
-        Registry registry = Registry.Builder.builder().withInsecure(true).build();
+        Registry registry = Registry.Builder.builder()
+                .withAuthProvider(authProvider)
+                .withInsecure(true)
+                .build();
 
         // Test
         List<String> tags = registry.getTags(ContainerRef.parse("%s/library/artifact-text"
+                .formatted(wmRuntimeInfo.getHttpBaseUrl().replace("http://", ""))));
+
+        // Assert
+        assertEquals(2, tags.size());
+        assertEquals("latest", tags.get(0));
+        assertEquals("0.1.1", tags.get(1));
+    }
+
+    @Test
+    void shouldListTagsWithFileStoreAuth(WireMockRuntimeInfo wmRuntimeInfo) throws IOException {
+
+        // Auth file for current registry
+        String authFile =
+                """
+                {
+                        "auths": {
+                                "localhost:%d": {
+                                        "auth": "bXl1c2VyOm15cGFzcw=="
+                                }
+                        }
+                }
+                """
+                        .formatted(wmRuntimeInfo.getHttpPort());
+
+        Files.writeString(configDir.resolve("config.json"), authFile, StandardCharsets.UTF_8);
+
+        ContainerRef containerRef = ContainerRef.forRegistry("localhost:%d".formatted(wmRuntimeInfo.getHttpPort()));
+        FileStoreAuthenticationProvider authProvider = new FileStoreAuthenticationProvider(
+                FileStore.newFileStore(configDir.resolve("config.json")), containerRef.getRegistry());
+
+        // Return data from wiremock
+        WireMock wireMock = wmRuntimeInfo.getWireMock();
+        wireMock.register(WireMock.get(WireMock.urlEqualTo("/v2/library/artifact-text-store/tags/list"))
+                .willReturn(WireMock.okJson(
+                        JsonUtils.toJson(new Tags("artifact-text-store", List.of("latest", "0.1.1"))))));
+
+        // Insecure registry
+        Registry registry = Registry.Builder.builder()
+                .withAuthProvider(authProvider)
+                .withInsecure(true)
+                .build();
+
+        // Test
+        List<String> tags = registry.getTags(ContainerRef.parse("%s/library/artifact-text-store"
                 .formatted(wmRuntimeInfo.getHttpBaseUrl().replace("http://", ""))));
 
         // Assert
